requirements.txt
django~=1.11
uwsgi
django-registration-redux
django-bootstrap3
django-markdownx
mistune
djangorestframework
django-filter
django-qr-code
pillow
